backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "phonenumbers>=9.0.14",
    "pydantic>=2.11.9",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.43",
    "types-python-jose>=3.5.0.20250531",
]

[dependency-groups]
dev = []
